Server: Create build script and update dependencies
This commit creates a Grunt file that contains the same tasks as the Web Client.
These include lint testing and a test runner.
The source files for server code has moved into the src directory and a main.js was created.
This should help with testing and making code more modular