Server: Build: Update jasmine test to start/stop mongoDB
This commit starts mongo then runs the tests. It will try and stop mongoDB
aswell but this might not work as grunt dies before running the stop mongo task.

Also this should not be run when the server or database is already running.
Some dependecies were added to help in testing and reading JSON from ajax.