Server: Add status property to User for login/logout feature
This commit adds a status option in the database which will be either
online, offline, or away. This allows the client to login and logoff.
Also there is now a cleanup for anonymous users that will delete the
user document on logout, when status changes to offline.

Also added a created data to user.